Move fmt
