+++ cli/src/config.rs
@@ -8,7 +8,7 @@ use color_eyre::Result;
 use eyre::{eyre, Report, WrapErr};
 use reqwest::header::USER_AGENT;
 use rink_core::context::Context;
-use rink_core::fmt::FmtToken;
+use rink_core::output::fmt::FmtToken;
 use rink_core::parsing::datetime;
 use rink_core::{ast, loader::gnu_units, CURRENCY_FILE, DATES_FILE, DEFAULT_FILE};
 use serde_derive::{Deserialize, Serialize};
