+++ core/src/output/fmt.rs
