+++ core/src/output/mod.rs
@@ -1,3 +1,4 @@
+pub mod fmt;
 mod reply;
 
 pub use reply::*;
